requirements.txt
aiohttp==3.9.1
openai==1.3.6
python-dotenv==1.0.0
replicate==0.21.0
streamlit==1.29.0
tenacity==8.2.2
